View the new flow and design
