+++ AngularApp/projects/app-service-diagnostics/src/app/diagnostic-tools/diagnostic-tools.routeconfig.ts
@@ -13,7 +13,6 @@ import { DaasMainComponent } from '../shared/components/daas-main/daas-main.comp
 import { Route } from '@angular/router';
 import { AutohealingDetectorComponent } from '../availability/detector-view/detectors/autohealing-detector/autohealing-detector.component';
 import { CpuMonitoringToolComponent } from '../shared/components/tools/cpu-monitoring-tool/cpu-monitoring-tool.component';
-import { MonitoringSessionsComponent } from '../shared/components/monitoring-sessions/monitoring-sessions.component';
 
 export const DiagnosticToolsRoutes: Route[] = [
     // CLR Profiling Tool
@@ -94,14 +93,6 @@ export const DiagnosticToolsRoutes: Route[] = [
         component: CpuMonitoringToolComponent,
         data: {
             navigationTitle: ToolNames.CpuMonitoring,
-            cacheComponent: true
-        }
-    },
-    {
-        path: 'cpumonitoringsessions',
-        component:MonitoringSessionsComponent,
-        data:{
-            navigationTitle:ToolNames.CpuMonitoringSessions,
             cacheComponent: true
         }
     },
