+++ AngularApp/projects/app-service-diagnostics/src/app/shared/components/daas/cpu-monitoring/cpu-monitoring-activity/cpu-monitoring-activity.component.scss
@@ -1,0 +1,3 @@
+.highlight-blue {
+    color: #286090;
+}
