+++ AngularApp/projects/app-service-diagnostics/src/app/shared/components/daas/cpu-monitoring/cpu-monitoring-configuration/cpu-monitoring-configuration.component.scss
@@ -1,0 +1,100 @@
+.label-header {
+    font-size: initial;
+    color:#0058ad;
+}
+
+.input-row {
+    min-height: 75px;
+}
+
+.tile-icon{
+    font-size: xx-large;
+    color:#0058ad;
+    min-width: 30px;
+}
+
+.summary-box { 
+    padding: 10px;
+    margin-left: 15px;
+    background-color: rgba(44, 163, 216, 0.1);
+    border:1px Solid #0058ad;
+}
+
+.summary-title{
+    margin-left: 15px;
+    background: #0058ad;
+    width: 180px;
+    border-top-left-radius: 3px;
+    border-top-right-radius: 3px;
+    padding: 2px;
+    padding-left:5px;
+    color:whitesmoke;
+}
+.media-body{
+    padding-left:13px;
+}
+
+::ng-deep {
+    .custom-slider .ng5-slider .ng5-slider-pointer:after {
+        width: 4px;
+        height: 4px;
+        top: 6px;
+        left: 6px;
+        border-radius: 4px;
+        }
+            
+    .custom-slider .ng5-slider .ng5-slider-pointer {
+        width: 16px;
+        height: 16px;
+        top: -6px;
+        border-radius: 13px;
+        margin-left: 4px;
+        background-color: #0058ad;
+    }
+}
+
+//https://codepen.io/triss90/pen/XNEdRe
+
+
+.radio {
+    margin: 0.5rem;
+  }
+  .radio input[type="radio"] {
+    position: absolute;
+    opacity: 0;
+  }
+  .radio input[type="radio"] + .radio-label:before {
+    content: '';
+    background: #f4f4f4;
+    border-radius: 100%;
+    border: 1px solid #b4b4b4;
+    display: inline-block;
+    width: 1.4em;
+    height: 1.4em;
+    position: relative;
+    top: -0.2em;
+    margin-right: 1em;
+    vertical-align: top;
+    cursor: pointer;
+    text-align: center;
+    transition: all 250ms ease;
+  }
+  .radio input[type="radio"]:checked + .radio-label:before {
+    // background-color: #3197EE;
+    background-color: #0058ad;
+    box-shadow: inset 0 0 0 4px #f4f4f4;
+  }
+  .radio input[type="radio"]:focus + .radio-label:before {
+    outline: none;
+    // border-color: #3197EE;
+    border-color: #0058ad;
+  }
+  .radio input[type="radio"]:disabled + .radio-label:before {
+    box-shadow: inset 0 0 0 4px #f4f4f4;
+    border-color: #b4b4b4;
+    background: #b4b4b4;
+  }
+  .radio input[type="radio"] + .radio-label:empty:before {
+    margin-right: 0;
+  }
+
