+++ AngularApp/projects/app-service-diagnostics/src/app/shared/components/daas/cpu-monitoring/cpu-monitoring-sessions/cpu-monitoring-sessions.component.html
@@ -1,0 +1,108 @@
+<div class="container">
+  <div class="row">
+      
+      <div class="col-sm-12">
+        <div *ngIf="activeSession">
+            <table class="table table-bordered">
+                <tr>
+                  <th>Start Date</th>
+                  <th *ngIf="activeSession.Mode !== 'Kill'">Logs</th>
+                  <th *ngIf="activeSession.Mode === 'CollectKillAndAnalyze'">Reports</th>
+                </tr>
+                <tr>
+                  <td>{{ formatDate(activeSession.StartDate)}}</td>
+                  <td *ngIf="activeSession.Mode !== 'Kill'">
+                    <div *ngIf="activeSession.FilesCollected == null || activeSession.FilesCollected.length === 0">
+                      No logs collected so far.
+                    </div>
+                    <ul *ngIf="activeSession.FilesCollected.length > 0" style="list-style-type:none;padding:0">
+                      <li *ngFor="let file of activeSession.FilesCollected">
+                          {{ getfileNameFromPath(file.RelativePath) }}
+                          <a (click)="openReport(file.RelativePath)">
+                              <i class="fa fa-external-link"></i>
+                          </a>
+                      </li>
+                    </ul>
+                  </td>
+                  <td *ngIf="activeSession.Mode === 'CollectKillAndAnalyze'">
+                    <div *ngIf="activeSession.AnalysisStatus === 'NotStarted'">
+                      Analysis will start after logs have been collected...
+                    </div>
+                    <div
+                      *ngIf="activeSession.AnalysisStatus !== 'NotStarted' && activeSession.AnalysisStatus !== 'Completed'">
+                      {{ activeSession.AnalysisStatus }}
+                    </div>
+                    <ul *ngIf="file.ReportFileRelativePath != null" style="list-style-type:none;padding:0">
+                      <li *ngFor="let file of activeSession.FilesCollected">
+                          {{ getfileNameFromPath(file.ReportFileRelativePath) }}
+                          <a (click)="openReport(file.ReportFileRelativePath)">
+                              <i class="fa fa-external-link"></i>
+                          </a>
+                      </li>
+                    </ul>
+                  </td>
+                </tr>
+              </table>
+        </div>
+        <div *ngIf="!activeSession">
+          No CPU Monitoring session active currently!
+        </div>
+      </div>
+  </div>
+  <div class="row">
+    <div class="col-sm-12">
+      <p>{{ description }}</p>
+      <div *ngIf="monitoringSessions.length > 0">
+        <table class="table table-hover table-bordered report-table">
+          <tr>
+            <th>Start Date</th>
+            <th>Session Duration</th>
+            <th>Logs</th>
+            <th>Reports</th>
+          </tr>
+          <tr *ngFor="let monitoringSession of monitoringSessions">
+            <td>
+              <div [innerHTML]="formatDate(monitoringSession.StartDate)"></div>
+            </td>
+            <td>
+                <div [innerHTML]="getDuration(monitoringSession.StartDate, monitoringSession.EndDate)"></div>
+              </td>
+            <td>
+              <div *ngIf="monitoringSession.FilesCollected == null || monitoringSession.FilesCollected.length === 0">
+                No logs collected.
+              </div>
+              <ul *ngIf="monitoringSession.FilesCollected.length > 0" style="list-style-type:none;padding:0">
+                <li *ngFor="let file of monitoringSession.FilesCollected">
+                    {{ getfileNameFromPath(file.RelativePath) }}
+                    <a (click)="openReport(file.RelativePath)">
+                        <i class="fa fa-external-link"></i>
+                    </a>
+                </li>
+              </ul>
+            </td>
+            <td>
+              <div
+                *ngIf="monitoringSession.AnalysisStatus !== 'NotStarted' && monitoringSession.AnalysisStatus !== 'Completed'">
+                {{ monitoringSession.AnalysisStatus }}
+              </div>
+              <button *ngIf="monitoringSession.AnalysisStatus === 'NotStarted' && monitoringSession.FilesCollected != null && monitoringSession.FilesCollected.length > 0"
+                (click)="analyzeSession(monitoringSession.SessionId)">Analyze</button>
+              <div *ngIf="monitoringSession.AnalysisStatus !== 'NotStarted'">
+                <ul style="list-style-type:none;padding:0">
+                  <li *ngFor="let file of monitoringSession.FilesCollected">
+                    <div *ngIf="file.ReportFileRelativePath != null">
+                      {{ getfileNameFromPath(file.ReportFileRelativePath) }}
+                      <a (click)="openReport(file.ReportFileRelativePath)">
+                          <i class="fa fa-external-link"></i>
+                      </a>
+                    </div>
+                  </li>
+                </ul>
+              </div>
+            </td>
+          </tr>
+        </table>
+      </div>
+    </div>
+  </div>
+</div>
