+++ AngularApp/projects/app-service-diagnostics/src/app/shared/components/daas/cpu-monitoring/cpu-monitoring-sessions/cpu-monitoring-sessions.component.scss
@@ -9,8 +9,3 @@
     padding-top: 10px;
     padding-bottom: 10px;
 }
-
-.container {
-    margin-left:20px;
-    margin-top:20px;
-}
