+++ AngularApp/projects/app-service-diagnostics/src/app/shared/components/daas/cpu-monitoring/cpu-monitoring.component.scss
@@ -5,14 +5,9 @@
 }
 
 .action-box {
-    // border: 1px solid #ebebeb;
     padding:15px;
-    //background-color: #f9f9f9
 }
 
-.highlight-blue {
-    color: #286090;
-}
 
 .mode-explanation{
     border: 0.5px solid #adb0b7;
@@ -28,36 +23,4 @@
 .mode-tab{
     border-bottom-right-radius:unset;
     border-bottom-left-radius: unset;
-}
-
-.summary-box {
-    background-color: #e4e4e4;
-    border-radius: 5px;
-    padding: 10px;
-    margin-left: 15px;
-    margin-top: 15px;
-    margin-bottom: 15px;
-}
-
-.monitoring-logs{
-    max-height: 300px;
-    overflow-y: scroll;
-}
-
-.label-header {
-    font-size: initial;
-    color:#0058ad;
-}
-
-.label-description {
-    font-size: small;
-    margin-top: 3px;
-}
-
-.input-row {
-    padding-top:8px;
-    padding-bottom: 15px;
-}
-i{
-    color:#0058ad;
 }
