--- AngularApp/projects/app-service-diagnostics/src/app/shared/components/monitoring-sessions/monitoring-sessions.component.html
@@ -1,37 +1,0 @@
-<div class="container">
-  <div class="header1">{{ title }}</div>
-  <p>{{ description }}</p>
-
-  <div *ngIf="monitoringSessions.length > 0">
-    <table class="table table-hover table-bordered report-table">
-      <tr *ngFor="let monitoringSession of monitoringSessions">
-        <td>
-          {{ monitoringSession.SessionId}}
-        </td>
-        <td>
-          <ul style="list-style-type:none">
-            <li *ngFor="let file of monitoringSession.FilesCollected">
-              <a (click)="openReport(file.RelativePath)">{{ getfileNameFromPath(file.RelativePath)}}</a>
-            </li>
-          </ul>
-        </td>
-        <td>
-          <div *ngIf="monitoringSession.AnalysisStatus !== 'NotStarted' && monitoringSession.AnalysisStatus !== 'Completed'">
-            {{ monitoringSession.AnalysisStatus }}
-          </div>
-          <button *ngIf="monitoringSession.AnalysisStatus === 'NotStarted'" (click)="analyzeSession(monitoringSession.SessionId)">Analyze</button>
-          <div *ngIf="monitoringSession.AnalysisStatus !== 'NotStarted'">
-            <ul style="list-style-type:none">
-              <li *ngFor="let file of monitoringSession.FilesCollected">
-                <div *ngIf="file.ReportFileRelativePath != null">
-                  <a (click)="openReport(file.ReportFileRelativePath)">{{
-                    getfileNameFromPath(file.ReportFileRelativePath) }}</a>
-                </div>
-              </li>
-            </ul>
-          </div>
-        </td>
-      </tr>
-    </table>
-  </div>
-</div>
