+++ AngularApp/projects/app-service-diagnostics/src/app/shared/components/number-picker/number-picker.component.scss
@@ -37,3 +37,7 @@
     text-align: center;
     width: 25px;
   }
+
+  input[type="text"]:disabled {
+    background: white;
+  }
