+++ AngularApp/projects/app-service-diagnostics/src/app/shared/shared.module.ts
@@ -86,10 +86,12 @@ import { GenericCommsService } from './services/generic-comms.service';
 import { GenericCommsComponent } from './components/generic-comms/generic-comms.component';
 import { LocalBackendService } from './services/local-backend.service';
 import { CpuMonitoringToolComponent } from './components/tools/cpu-monitoring-tool/cpu-monitoring-tool.component';
-import { MonitoringSessionsComponent } from './components/monitoring-sessions/monitoring-sessions.component';
 import { CpuMonitoringComponent } from './components/daas/cpu-monitoring/cpu-monitoring.component';
 import { NumberPickerComponent } from './components/number-picker/number-picker.component';
 import { Ng5SliderModule } from 'ng5-slider';
+import { CpuMonitoringConfigurationComponent } from './components/daas/cpu-monitoring/cpu-monitoring-configuration/cpu-monitoring-configuration.component';
+import { CpuMonitoringActivityComponent } from './components/daas/cpu-monitoring/cpu-monitoring-activity/cpu-monitoring-activity.component';
+import { CpuMonitoringSessionsComponent } from './components/daas/cpu-monitoring/cpu-monitoring-sessions/cpu-monitoring-sessions.component';
 
 @NgModule({
     declarations: [
@@ -149,9 +151,11 @@ import { Ng5SliderModule } from 'ng5-slider';
         TileListComponent,
         GenericCommsComponent,
         CpuMonitoringToolComponent,
-        MonitoringSessionsComponent,
         CpuMonitoringComponent,
-        NumberPickerComponent
+        NumberPickerComponent,
+        CpuMonitoringConfigurationComponent,
+        CpuMonitoringActivityComponent,
+        CpuMonitoringSessionsComponent
     ],
     imports: [
         HttpModule,
